Mise en place de l'administration dans AdminEventController
